// src/interfaces/rh/RhTmpRetencionesCahDto.ts
export interface IRhTmpRetencionesCahDto {
  id: number
  codigoRetencionAporte: number
  secuencia: number
  unidadEjecutora: string
  cedulaTexto: string
  nombresApellidos: string
  descripcionCargo: string
  fechaIngreso: Date
  montoCahTrabajador: number
  montoCahPatrono: number
  montoTotalRetencion: number
  fechaNomina: Date
  siglasTipoNomina: string
  fechaDesde: Date
  fechaHasta: Date
  codigoTipoNomina: number
  searchText: string
  avatar: string
}